System Prefs: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@39271 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/prefs/font/fpeditor.c
@@ -43,7 +43,7 @@ static VOID FontPrefs2FontString
     snprintf
     (
         buffer, buffersize, "%.*s/%d",
-        strlen(fp->fp_TextAttr.ta_Name) - 5 /* strlen(".font") */,
+        (int)strlen(fp->fp_TextAttr.ta_Name) - 5 /* strlen(".font") */,
         fp->fp_TextAttr.ta_Name, fp->fp_TextAttr.ta_YSize
     );
 }
